Rename `ShortCodeValidationResult` to `EnteredCodeValidationResult`
