+++ app/src/main/java/org/simple/clinic/scanid/EnteredCodeInput.kt
@@ -4,9 +4,9 @@ import android.os.Parcelable
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BpPassport.SHORT_CODE_LENGTH
-import org.simple.clinic.scanid.ShortCodeValidationResult.Failure.Empty
-import org.simple.clinic.scanid.ShortCodeValidationResult.Failure.NotEqualToRequiredLength
-import org.simple.clinic.scanid.ShortCodeValidationResult.Success
+import org.simple.clinic.scanid.EnteredCodeValidationResult.Failure.Empty
+import org.simple.clinic.scanid.EnteredCodeValidationResult.Failure.NotEqualToRequiredLength
+import org.simple.clinic.scanid.EnteredCodeValidationResult.Success
 
 @Parcelize
 data class EnteredCodeInput(val shortCodeText: String) : Parcelable {
@@ -14,7 +14,7 @@ data class EnteredCodeInput(val shortCodeText: String) : Parcelable {
   @IgnoredOnParcel
   private val longCodeLength = 14
 
-  fun validate(): ShortCodeValidationResult {
+  fun validate(): EnteredCodeValidationResult {
     return when {
       shortCodeText.isEmpty() -> Empty
       shortCodeText.length != SHORT_CODE_LENGTH && shortCodeText.length != longCodeLength -> NotEqualToRequiredLength
@@ -24,10 +24,10 @@ data class EnteredCodeInput(val shortCodeText: String) : Parcelable {
   }
 }
 
-sealed class ShortCodeValidationResult {
-  object Success : ShortCodeValidationResult()
+sealed class EnteredCodeValidationResult {
+  object Success : EnteredCodeValidationResult()
 
-  sealed class Failure : ShortCodeValidationResult() {
+  sealed class Failure : EnteredCodeValidationResult() {
     object Empty : Failure()
     object NotEqualToRequiredLength : Failure()
   }
