+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdUiActions.kt
@@ -4,6 +4,6 @@ interface ScanSimpleIdUiActions {
   fun showQrCodeScannerView()
   fun hideQrCodeScannerView()
   fun hideShortCodeValidationError()
-  fun showShortCodeValidationError(failure: ShortCodeValidationResult)
+  fun showShortCodeValidationError(failure: EnteredCodeValidationResult)
   fun sendScannedId(scanResult: ScanResult)
 }
