+++ app/src/test/java/org/simple/clinic/scanid/ScanSimpleIdUpdateTest.kt
@@ -47,7 +47,7 @@ class ScanSimpleIdUpdateTest {
 
     spec
         .given(model)
-        .whenEvent(ShortCodeValidated(ShortCodeValidationResult.Success))
+        .whenEvent(ShortCodeValidated(EnteredCodeValidationResult.Success))
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(SendScannedIdentifierResult(expectedScanResult))
